Bug fix for not restore volume level
There appears to be a bug in the DTH while handling volume requests. In `setVolumeSpeakAndRestore` it saves the current volume in `state.lastVolume` after reading it from `level`, the new volume in `state.useThisVolume` and updates the attribute `level` to the new volume. Now it calls `speak` which uses state.useThisVolume and `state.lastVolume` to make the call to Echo which set the new volume, speaks and restore the old volume on the Echo (correctly).

However when it's done with the command it doesn't restore the attribute `level` back to the `state.lastVolume` (this it the bug). So the next time `setVolumeSpeakAndRestore` is called, it read the attribute `level`, which wasn't restore so it contains the previous volume so it instead of using the restore volume now the new volume is the previously requested volume.

From what I see restoreLastVolume won't work because `queueEchoCmd` sets `state.lastVolume` to null.

It also fixes an issue where the `use_cnt_setVolumeSpeakRestore` isn't incremented if the volume level isn't changed